Updated portfolio code
I've made some improvements to the typography: added a CSS grid for responsive layout. The section with the latest notes is now extracted into a separate blog: you can add a link to a new note and add a photo to it. Links are graphically represented and underlined with animation.

On the main page, there are now three link-buttons: one leads to Elena's LinkedIn profile, and the other to her GitHub. I've also added a button to the footer that, when clicked, copies Elena's email to the clipboard.